Javadoc fully sync'ed with  JDK 1.7 M11 2010/09/10 (b123)
TODO:  Review and implement the JDBC 4.1 behavioral changes specified in the Javadoc updates.

1.) add IS_GENERATED column to getColumns result.
2.) ensure updated result column ordering specifications hold and update/future-proof internal code where they do not currently hold or at least may not hold in the future.
3.) ensure that updated result column value specifications hold.  In particular, ensure that NULL column values are generated where JDBC 4.1 has updated the expected value from [0] to [NULL]

Review whether point 2 and 3 should be handled as blanket retroactive [were incorrectly specified in previous JDBC versions] or need to follow a 'cutover' policy, i.e. need to test if the product was built is being run under JDK 7+, a.k.a is implementing JDBC version 4.1, not an earlier specification version.

Most likely, point 2 is a clarification and as such is blanket retroactive.  Regardless, we still really don't need to work about it, as we don't support multiple catalogs yet.  However, probably a good idea to future-proof the underlying code by updating the primary index specs for the generating the corresponding information schema system tables.

On the other hand, 3 must be reviewed a bit more carefully.  

If it really means that the expectation all along has been for the driver to return [0] when getShort(...), getInt(...), getLong(...) etc. was called on the ResultSet (and wasNull() would return true), then this is a clarification (that getObject(...)  should be expected to return NULL).
If not a clarification, then it is potentially a breaking change for clients who invoke getObject asking for an allowable java object mapping and expect a non-null return value.